refactor(#149): ImageGallery 컴포넌트 폴더 이동
