+++ client/src/components/ImageGallery.tsx
